refactor: separated metadata concerns from audio
