+++ src-tauri/src/audio/mod.rs
@@ -1,0 +1,4 @@
+pub mod backend;
+pub mod commands;
+pub mod state;
+pub mod utils;
