+++ src-tauri/src/metadata/mod.rs
@@ -1,0 +1,3 @@
+pub mod commands;
+pub mod metadata;
+pub mod utils;
